% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{estimation_error}
\alias{estimation_error}
\title{evaluate estimation error}
\usage{
estimation_error(true_pc, estimated_pc)
}
\arguments{
\item{true_pc}{true principal component}

\item{estimated_pc}{estimated principal component}
}
\value{
\describe{
\item{alignment}{a float between 0 and 1, alignment between true and estimated PC}
}
}
\description{
evaluate estimation error
}
