% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGPCA.R
\name{SGPCA}
\alias{SGPCA}
\title{Sparse Group Principal Component Analysis for multiple PCs estimation}
\usage{
SGPCA(
  X,
  group_label,
  J = 1,
  tau_reg = 1,
  eta_reg = 1,
  max_iter = 100,
  tol = 1e-05
)
}
\arguments{
\item{X}{a data matrix}

\item{group_label}{a vector for group membership labels}

\item{J}{number of principal components to be estimated}

\item{tau_reg}{a float to specify the individual thresholding level}

\item{eta_reg}{a float to specify the group thresholding level}

\item{max_iter}{a positive integer to specify the maximum iterations to operate}

\item{tol}{a float to specify the stopping threshold}
}
\value{
a list of the following components
\describe{
\item{V}{a P * J matrix, estimated J PCs}
\item{U}{a N * J matrix, scores on the estimated J PCs}
\item{D}{a J-dimensional vector, variance explained by J PCs}
}
}
\description{
Sparse Group Principal Component Analysis for multiple PCs estimation
}
