% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGPCA.R
\name{SGPCA.rs}
\alias{SGPCA.rs}
\title{Tuning parameter for SGPCA using resampling-based method and alignment criterion}
\usage{
SGPCA.rs(
  X,
  group_label,
  J = 1,
  B = 20,
  rho = 0.5,
  tau_range = 10^seq(-2, 2, length.out = 20),
  eta_range = 10^seq(-2, 2, length.out = 20),
  max_iter = 100,
  tol = 1e-05,
  verbose = TRUE,
  mode = "auto"
)
}
\arguments{
\item{X}{a data matrix}

\item{group_label}{a vector for group membership labels}

\item{J}{number of principal components to be estimated}

\item{B}{number of resampling datasets}

\item{rho}{resampling ratio}

\item{tau_range}{search range for tau}

\item{eta_range}{search range for eta}

\item{max_iter}{a positive integer to specify the maximum iterations to operate}

\item{tol}{a float to specify the stopping threshold}

\item{verbose}{TRUE/FALSE, output algorithm info (TRUE) or not (FALSE)}

\item{mode}{string, "auto" or "manual". "auto" mode return the tuning parameter pairs that maximize the alignment score. "manual" mode output a figure "alignment v.s. mean support size", user can specify the tuning parameters interactively.}
}
\value{
a list of the following components
\describe{
\item{V}{a P * J matrix, estimated J PCs}
\item{U}{a N * J matrix, scores on the estimated J PCs}
\item{D}{a J-dimensional vector, variance explained by J PCs}
}
}
\description{
Tuning parameter for SGPCA using resampling-based method and alignment criterion
}
