% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGPCA.R
\name{SGPCA_1pc}
\alias{SGPCA_1pc}
\title{Sparse Group Principal Component Analysis for one single PC estimation}
\usage{
SGPCA_1pc(
  X,
  group_label,
  tau_reg = 1,
  eta_reg = 1,
  max_iter = 100,
  tol = 1e-05
)
}
\arguments{
\item{X}{a data matrix}

\item{group_label}{a vector for group membership labels}

\item{tau_reg}{a float to specify the individual thresholding level}

\item{eta_reg}{a float to specify the group thresholding level}

\item{max_iter}{a positive integer to specify the maximum iterations to operate}

\item{tol}{a float to specify the stopping threshold}
}
\value{
A list of with the following components:
\describe{
\item{v}{a P-dimensional vector, estimated the first PC}
\item{u}{a N-dimensional vector, scores of the first PC}
\item{d}{a float, variance explained by the first PC}
}
}
\description{
Sparse Group Principal Component Analysis for one single PC estimation
}
