% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator2}
\alias{simulator2}
\title{Generator of simulation setting 2}
\usage{
simulator2(n = 500, G = 2000, cond = 3, seed = 123)
}
\arguments{
\item{n}{sample size}

\item{G}{number of genes}

\item{cond}{sprsity condition (1: group sparsity, 2: individual sparsity, 3: group and individual sparsity)}

\item{seed}{an integer to specify the random seed}
}
\value{
\describe{
\item{X}{data matrix of dimension n by (10G)}
\item{group_label}{a (10G)-dimensional vector, group membership}
\item{signal_indices}{a vector, provides the indices of signals in PC}
\item{pc1}{a (10G)-dimensional vector, true PC1 by design}
}
}
\description{
Generator of simulation setting 2
}
