% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator1}
\alias{simulator1}
\title{Generator of simulation setting 1}
\usage{
simulator1(n = 100, G = 300, C = 3, seed = 123)
}
\arguments{
\item{n}{sample size}

\item{G}{number of genes}

\item{C}{group size (e.g., number of celltypes)}

\item{seed}{an integer to specify the random seed}
}
\value{
\describe{
\item{X}{data matrix of dimension n by (G * C)}
\item{group_label}{a (G * C)-dimensional vector, group membership}
\item{signal_indices}{a vector, provides the indices of signals in PC}
\item{pc1}{a (G * C)-dimensional vector, true PC1 by design}
}
}
\description{
Generator of simulation setting 1
}
